+++ tests/test_generation.py
@@ -5,7 +5,7 @@ from flourish import Flourish
 from .compare_directories import CompareDirectories
 
 
-class TestFlourishGeneration(CompareDirectories):
+class FullGeneration(CompareDirectories):
     expected_directory = 'tests/output'
     expected_files = [
         '2015/12/25/index.html',
@@ -32,15 +32,10 @@ class TestFlourishGeneration(CompareDirectories):
         'markdown-page.html',
         'nope.txt',
         'nothing.html',
-        'thing-one.html',
-        'thing-two.html',
         'series/index.html',
         'series/part-one.html',
         'series/part-three.html',
         'series/part-two.html',
-        'tags/index/index.atom',
-        'tags/index/index.html',
-        'tags/index/series/index.html',
         'tags/basic-page/basic-page.html',
         'tags/basic-page/index.atom',
         'tags/basic-page/index.html',
@@ -51,6 +46,9 @@ class TestFlourishGeneration(CompareDirectories):
         'tags/first/index.atom',
         'tags/first/index.html',
         'tags/first/thing-one.html',
+        'tags/index/index.atom',
+        'tags/index/index.html',
+        'tags/index/series/index.html',
         'tags/one/index.atom',
         'tags/one/index.html',
         'tags/one/series/part-one.html',
@@ -62,8 +60,8 @@ class TestFlourishGeneration(CompareDirectories):
         'tags/series/index.html',
         'tags/series/series/index.html',
         'tags/series/series/part-one.html',
-        'tags/series/series/part-two.html',
         'tags/series/series/part-three.html',
+        'tags/series/series/part-two.html',
         'tags/three/index.atom',
         'tags/three/index.html',
         'tags/three/series/part-three.html',
@@ -71,13 +69,37 @@ class TestFlourishGeneration(CompareDirectories):
         'tags/two/index.html',
         'tags/two/series/part-two.html',
         'tags/two/thing-two.html',
+        'thing-one.html',
+        'thing-two.html',
     ]
 
+
+class TestFlourishGeneration(FullGeneration):
     def test_generation(self):
         with pytest.warns(None) as warnings:
             flourish = Flourish(
                 source_dir='tests/source',
                 templates_dir='tests/templates',
+                sass_dir='tests/sass',
+                output_dir=self.tempdir,
+            )
+            assert len(warnings) == 2
+
+        with pytest.warns(None) as warnings:
+            # one template has an invalid url() use
+            flourish.generate_site()
+            assert len(warnings) == 1
+            assert 'tags-tag-page' in str(warnings[0].message)
+
+        self.compare_directories()
+
+
+class TestSectileTemplatesGeneration(FullGeneration):
+    def test_generation(self):
+        with pytest.warns(None) as warnings:
+            flourish = Flourish(
+                source_dir='tests/source',
+                fragments_dir='tests/fragments',
                 sass_dir='tests/sass',
                 output_dir=self.tempdir,
             )